+++ packages/ramps-controller/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add sync trigger methods to RampsController ([#7662](https://github.com/MetaMask/core/pull/7662))
+
+- Export `RampAction` type for `'buy' | 'sell'` ramp actions ([#7663](https://github.com/MetaMask/core/pull/7663))
+
 ## [4.0.0]
 
 ### Added